// src/enums/ColumnsActions.ts
export enum ColumnsActions {
    DocumentType = 'DocumentType',
    DocumentStatus = 'DocumentStatus',
    Notification = 'Notification',
    ResultStatus = 'ResultStatus',
    Publish = 'Publish',
    AllowComment = 'AllowComment',
    ActionGroups = 'ActionGroups',
    Phase = 'Phase',
    NotificationGroups = 'NotificationGroups',
    Task = 'Task',
    TaskGroups = 'TaskGroups',
    TitleEn = 'Title_en',
    NotificationEn = 'Notification_en',
    TaskEn = 'Task_en',
    ReadPermissions = 'ReadPermissions',
    WritePermissions = 'WritePermissions',
    Validity = 'validity',
    Reupload = 'Reupload',
    FileLink = 'FileLink',
    OnMajorVersion = 'onMajorVersion'
}